requirements.txt
# Core ML dependencies (CPU-only for faster deployment on Render)
--index-url https://download.pytorch.org/whl/cpu
torch==2.1.0
torchvision==0.16.0

# Other dependencies
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.1.0

# FastAPI dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
